requirements.txt
numpy
nltk
pdfplumber
sentence-transformers
faiss-cpu
